refactor(getCountryInfo) : getCountryInfo -> country로 이름변경
